+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **breaking change** Harvest backend is now sync [#3030](https://github.com/opendatateam/udata/pull/3030)
 - Allow dataservices to be discussed and followed [#3049](https://github.com/opendatateam/udata/pull/3049)
+- Add purge-dataservices job [#3049](https://github.com/opendatateam/udata/pull/3049)
 
 ## 8.0.1 (2024-05-28)