+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **breaking change** Harvest backend is now sync [#3030](https://github.com/opendatateam/udata/pull/3030)
 - Allow dataservices to be discussed and followed [#3049](https://github.com/opendatateam/udata/pull/3049)
 - Add purge-dataservices job [#3049](https://github.com/opendatateam/udata/pull/3049)
+- Harvest all the available polygons from a spatial coverage [#3039](https://github.com/opendatateam/udata/pull/3039)
 
 ## 8.0.1 (2024-05-28)